refactor(mobile): extract queue provider session commands
